Reworked to include economy in only one extension as requested
